Fix serial data line for pll
PLL does not seem to lock, but with the discovery of the typo for the
pin used for the data, it seems that we can now use the SPI1 peripheral
to send the data, and with a faster clock rather than bitbanging. Maybe
that will make it work?

Also I note from the spreadsheet that 32 bit words are sent via the SPI
and not the 19 bite as required by the PLL.